requirements.txt
requests
sodapy
pandas
numpy
geopandas
fiona
shapely
pyarrow
pdfplumber
psycopg2-binary
sqlalchemy
python-dotenv
apscheduler
scikit-learn
xgboost
openpyxl
